Revive this repo with new examples
In this commit we are resurrecting this repository from the dead. Long
ago, when opera was still very young this repository was established to
hold complex opera examples. Those were meant to be separated as if the
/examples folder in this repository are about "how do I do a specific
thing", then the separate repository is more about "what are all the
wonderful options that TOSCA and opera opens up for me". Now after some
time we decided to make use of this old repository. Since the old
examples were still using TOSCA YAML v1.0, we had to transform them
into v1.3. We then collected some available examples that are
deployable with opera and added them here so that this will now be a
central point for all opera examples.