Establish CI validation for examples
Here we are adding a new GitHub Actions CI configuration that will
validate all the examples within this repository. Thereby we will be
able to be sure that all the examples are deployable with xOpera
orchestrator. This CI job will be triggered on every push event and
will test the examples with the latest opera production version by
running a prepared bash script. The examples are validated using opera
validate CLI command. We are not currently testing deploy, undeploy,
update and other possible commands on those examples because we rely on
the creators that the prepared examples work and the other reason is
that the deployment of all example would consume a lot of time and also
some examples cannot be directly deployed since they require some
special prerequisites.