Add another folder with TOSCA CSAR examples
After supplying initial xOpera examples and establishing the desired
skeleton of this repository, we are now adding a folder that will
contain the extracted TOSCA CSAR examples. Although TOSCA CSARs should
be originally compressed files, we decided to keep them extracted since
it's not a good practice to keep compressed archives in a Git
repository. The compressed version of those CSARs may end up in our new
solution for storing TOSCA entities that we're building (called
Template Library).