Comment some lines in validation script due to opera CSAR bug
The current opera version 0.6.4 has a bug and works only with CSARs
that have TOSCA.meta file and does not work with the CSARs that have
metadata specified in TOSCA service template. When new opera will be
released we will be able to uncomment these lines so that the small
CSAR example can be fully tested.