Supply CSAR examples (#7)
* Add another folder with TOSCA CSAR examples

After supplying initial xOpera examples and establishing the desired
skeleton of this repository, we are now adding a folder that will
contain the extracted TOSCA CSAR examples. Although TOSCA CSARs should
be originally compressed files, we decided to keep them extracted since
it's not a good practice to keep compressed archives in a Git
repository. The compressed version of those CSARs may end up in our new
solution for storing TOSCA entities that we're building (called
Template Library).

* Supply README files for subfolders

Our xOpera examples are separated into folders based on their target
usage scenario. To describe each of these scenarios we added README
files to all the folders that represent different types of examples.

* Correct the contributors badge link

This badge link was corrected because it was pointing to the opera CLI
repository instead of this one.

* Comment some lines in validation script due to opera CSAR bug

The current opera version 0.6.4 has a bug and works only with CSARs
that have TOSCA.meta file and does not work with the CSARs that have
metadata specified in TOSCA service template. When new opera will be
released we will be able to uncomment these lines so that the small
CSAR example can be fully tested.