Remove requirements.txt and update CI validation
Within this commit we are removing the requirements.txt since opera
examples only need opera pip package to validate them. This commit was
first meant to use the latest opera dev version for validation and to
apply some changes that would let use use the latest opera pre-release
version from TestPyPI for validation GitHub Action workflow. This would
enable using the latest opera features within examples immediately as
new opera dev pre-release is made every time when new commits are
pushed on master branch. For the nightly tests that are yet to be
implemented, we can use the production opera version to ensure the
stability of opera examples. Or the other way around - we can use opera
production version for regular validation and dev version for nightly
testing. But we will do this in our future PRs.